+++ R/plot_scatterplot.r
@@ -3,6 +3,7 @@
 #' This function creates scatterplot for all features fixing on a selected feature.
 #' @param data input data
 #' @param by feature name to be fixed at
+#' @param sampled_rows number of rows to sample if data has too many rows. Default is all rows, which means do not sample.
 #' @param geom_point_args a list of other arguments to \link{geom_point}
 #' @param title plot title
 #' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
@@ -28,11 +29,13 @@
 #'   ncol = 4L
 #' )
 
-plot_scatterplot <- function(data, by, geom_point_args = list(), title = NULL, ggtheme = theme_gray(), theme_config = list(), nrow = 3L, ncol = 3L) {
+plot_scatterplot <- function(data, by, sampled_rows = nrow(data), geom_point_args = list(), title = NULL, ggtheme = theme_gray(), theme_config = list(), nrow = 3L, ncol = 3L) {
 	## Declare variable first to pass R CMD check
 	variable <- NULL
 	## Check if input is data.table
 	if (!is.data.table(data)) data <- data.table(data)
+	## Sample data if necessary
+	if (sampled_rows < nrow(data)) data <- data[sample.int(nrow(data), sampled_rows)]
 	## Create plot function
 	dt <- suppressWarnings(melt.data.table(data, id.vars = by, variable.factor = FALSE))
 	feature_names <- unique(dt[["variable"]])