+++ tests/testthat/test-plot-histogram-density.r
@@ -9,11 +9,21 @@ test_that("test column names containing spaces", {
 test_that("test return object", {
 	histogram_list <- plot_histogram(iris)
 	expect_is(histogram_list, "list")
-	expect_equal(names(histogram_list), c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))
-	expect_true(all(vapply(histogram_list, is.ggplot, TRUE)))
+	expect_equal(names(histogram_list), "page_1")
+	expect_true(is.ggplot(histogram_list[[1]]))
+
+	histogram_list2 <- plot_histogram(iris, nrow = 1L, ncol = 1L)
+	expect_is(histogram_list2, "list")
+	expect_equal(names(histogram_list2), c("page_1", "page_2", "page_3", "page_4"))
+	expect_true(all(vapply(histogram_list2, is.ggplot, TRUE)))
 
 	density_list <- plot_density(iris)
 	expect_is(density_list, "list")
-	expect_equal(names(density_list), c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))
-	expect_true(all(vapply(density_list, is.ggplot, TRUE)))
+	expect_equal(names(density_list), "page_1")
+	expect_true(is.ggplot(density_list[[1]]))
+
+	density_list2 <- plot_density(iris, nrow = 1L, ncol = 1L)
+	expect_is(density_list2, "list")
+	expect_equal(names(density_list2), c("page_1", "page_2", "page_3", "page_4"))
+	expect_true(all(vapply(density_list2, is.ggplot, TRUE)))
 })