+++ NEWS.md
@@ -11,7 +11,10 @@
 * [#78](https://github.com/boxuancui/DataExplorer/issues/78): All ggplot objects from `plot_*` are now invisibly returned. As a result, extracted `profile_missing` from `plot_missing` for missing value profiles.
 * [#83](https://github.com/boxuancui/DataExplorer/issues/83): Removed all deprecated functions.
 * [#85](https://github.com/boxuancui/DataExplorer/issues/85): Users can now specify number of rows/columns for plot page layout.
+* `plot_prcomp` now passed `scale. = TRUE` to `prcomp` by default.
+* Added `sampled_rows` argument to `plot_scatterplot`.
 * Plot objects are now constructed with `parallel::mclapply`.
+* Updated default config for `create_report`.
 
 #### Bug Fixes
 * [#74](https://github.com/boxuancui/DataExplorer/issues/74): Fixed a bug causing `create_report` failure due to zero complete rows.